Re-implement change of IO::Handle encoding.
Allowing a type object to be passed to indicate "no encoding", and
making an argument of 'bin' translate into that.

The most common use for this is a switch of encoding prior to doing
any I/O, but when passing it at `open` time is not an option. For all
other cases, we try to cope. Binary mode to having an encoding is the
straightforward case. Char mode to binary mode needs to "un-read" any
content in the decode buffer, and for now will only be allowed on a
handle that supports `seek`. Char mode to char mode with a different
encoding takes the bytes unused by the previous decoder and gives them
to the new one. In both of these cases, it's only generally safe to do
this after a control character (like a newline), otherwise bytes may
be hanging around inside of decode/normalization buffers if switching
away from, say, UTF-8. (This is not a change of behavior; in fact, we
allow *more* than we used to, given we get a couple of newly passing
tests.)