Update IO::Path.slurp for IO::Handle changes.
This still tries to do some amount of fast-pathing by avoidng the cost
of the setup of a streaming decoder by opening in binary mode (or at
least, we'll do so once MoarVM's file handle support switches to be
binary only), and doing the decode in a single swoop. More may be
possible; in the meantime, this gets things working again.