Revert "Revert "Close unit's file handle after reading in the bytecode""
This reverts commit f027779177593ea5bf07d3ac7ee06591af0d7946.

Zoffix++'s fix may actually have been correct after all.
From mmap (3p): "The mmap() function shall add an extra reference to the file
associated with the file descriptor fildes which is not removed by a subsequent
close() on that file descriptor. This reference shall be removed when there
are no more mappings to the file."

As I don't see any alternative fix anyway, let's give this an honest try.