[io grant] Remove IO::Path.watch
There are zero tests for it and the whole notifications business
is quite rough on the edges and probably needs more work. Remove
the method so we don't lock ourselves into supporting some API
that might not be suitable once notifications are sorted out.

The functionality is still available via
IO::Notification.watch-path: $path