Tweak X::Method::NotFound some more
Scale the max string distance allowed for suggested methods by the
length of the non-existant method.

Use the :all parameter to ^methods() to pick up possible suggestions
from Mu, Any, Cool, etc..

Merge the hard-coded suggestions for 'length' and 'bytes' in with the
string difference suggestions, but with a 0 length, so they'll always be
suggested.

Store the distance for all methods and then select the lowest four to
suggest.