Fix hang in deepmap with Iterable type objects
If an Iterable type object is given to deepmap, .list gets called
on it and the resultant 1-el iterable with the type object inside
of it gets used to map over. However, the logic that descends into
iterables forgets to check if they're type objects, so we end up
infinitely recursing by descending into the type object iterable,
calling .list on it, etc.

Fix by adding a check for .DEFINITEness before descending.

Bug find: https://irclog.perlgeek.de/perl6/2017-06-01#i_14672468