Fix crash when Junctioning empty array after .elems'ing it
`my @a; say +@a; say @a.all` crashes with
`This type (Scalar) does not support elems` because the `.elems`
call touches $!reified attr, making it nqp::attrinited.
Then, when we go to make a shiny new Junction, we check if the
$!reified is inited and use it if it is. However, it's inited because
.elems touched it, not because it got stuff up in it.

Fix by checking concreteness of $!reified instead of initness.