Give Proc::Async support for merged output.
Available through the `.Supply` method, which feels fairly natural.
In IO::Socket::Async, .Supply is how you get the data received over
the socket; `.Supply` on a process can be seen as "get me the output"
without caring which output.

Added as part of work towards re-working Proc in terms of Proc::Async,
which means giving Proc::Async the features that Proc has. However, as
it turns out :merge was busted and disabled for Proc, this counts as a
bonus, I guess.