Don't show "Actually thrown at:" info if duplicate
This was bugging me for a while.  We'd lose the fact that it was a
Failure that was thrown lazily, but *only* if it was thrown at the
same location it was received.  And with this fix, testing one-liners
gets a lot shorter if a Failure was involved.

Breaks one spectest that can be fixed easily.  Please revert if
considered too magical or for any other reason :-)