A worry for RT #126732
I blinked my eyes and the ticket got closed. In any case, here is a
fix.

For those who think that this might cause some kind of breakage, this
is just a worry, so don't worry.

This whole discussion was basically about adding or not adding this
one line of code. I don't see any problem with it. Curious users will
try to use superscripts alone, maybe just for fun, but a “potential
difficulty” will give them a lesser WTF-ish feeling. So let's just
take it.