Introduce read-internal and write-internal.
These are working names for the operations. They are the low-level,
internal, methods for reading and writing from a handle. Since all
of the higher-level ones now call them, this means that it is much
easier to subclass IO::Handle: only this pair of methods and a
handful of others need be overridden, and then you get all of the
string-level and mixed-mode input and string-level output for free.

This is in preparation for IO::Pipe switching to use Proc::Async,
which will become a concrete illustration of this.