Change the way we code-gen simple for loops.
For a for loop with a single argument and no phasers, we can generate
inline code that calls .pull-one and invokes the block. This will turn
megamorphic callsites inside of sink-all into typically monomorphic
ones, which should enable the specializer - at least, once it has had
some tweaks - to do a better job.