Fix ModuleLoader.nqp so RAKUDO_MODULE_DEBUG works the same as elsewhere
Fixes a bug that would cause Rakudo to die during compilation
if RAKUDO_MODULE_DEBUG were set to anything other than a number. This was
discovered by the Rakudo AppImage builder here:
https://travis-ci.org/samcv/rakudo-appimage-module-test-automation/jobs/242732276#L1204

When given a non-number during normal operation the debugging code in
ModuleLoader.nqp would not be activated, even though other debugging
code in Rakudo codebase would. The original code was added a long time ago,
on 2013-04-21.

This fix only applies to MoarVM, and this bug could still exist on the JVM.
This commit harmonizes the ModuleLoader.nqp code with the Perl 6 based code.

$DEBUG is set to 1 for Truey numeric values of ENV 'RAKUDO_MODULE_DEBUG'
or for non-numeric strings. All other cases $DEBUG is set to 0