Unbust the perl6-debug-m debugger
The .define_slang stuff seems to be what's busting it. Unsure why.
The breakage appears to occur whenever strings are used (so something
to do with the switch to Quote lang?). For example -e 'say 42' won't
crash, but -e 'say "x"' will. Even -e '"x"' crashes. Looking at the
AST of the last one, shows (NQPMu) in the tree, where the working
debugger's AST shows a QAST::Want for the WVal(Str)/SVal.