[unicode-grant] Collation object updated to work with new unicmp_s semantics
The Collation object now allows setting reverse sort order of different
collation levels. This is backwards compatible as long as the user
has not manually set the collation-level themselves. This was strongly
warned against relying on for this experimental API.

Setting the collation levels the recommended way still yields the same
result. In addition you can now set the collation levels to negative
values to reverse the sorting order for just that collation level.