Add Encoding and Encoding::Encoder roles.
Where Encoding represents an encoding, and can be used to obtain both
a streaming decoder and an encoder. The Encoding::Encoder role
represents an encoder (which is, thankfully, a far simpler thing than
a decoder).

It's tempting to call the method just `encode`, but given it will have
a different API to `Str.encode` it may be better to have it different
(a `Encoding::Encoder` will pre-commit to things like replacement
characters and newline translation).