Move Rakudo::Internals::VMBackedDecoder.
It's now `Encoding::Decoder::Builtin`. We'll not spec or document this
type, but rather say that for built-in encodings the exact types you
get back from `Encoding::Registry.find`, and from `.decoder` and
`.encoder` on the retunred object, are all implementation defined and
so subject to change; the only thing to be relied on is them doing the
`Encoding`/`Encoding::Encoder`/`Encoding::Decoder` roles.

Also generalize various places to the `Encoding::Decoder` role.