Add a `buffer` argument to IO::Handle.
When passed a boolean, it either enables or disables output buffering.
Anything else will be coerced to an `Int`. So, `:buffer` means buffer
output with a default buffer size, `:!buffer` means do not buffer
output, and `:4096buffer` means to use a buffer size of 4096 bytes.

For now the default is `:!buffer`; in the future, we'll try changing
it to be `:buffer` in the case that the file handle is not a TTY,
and beyond that enable it for the standard handles in non-TTY mode
also.