Speed up starting installed scripts
Turns out we don't actually have to run a separate perl6 process just to run
the appropriate installed script from the wrapper. Due to insufficient checks
in Perl6::Actions, MAIN subs are even run when loading a script with
"require". Arguably that's a bug, but for now this is useful.

If we just changed the created wrapper scripts to use "require", we'd have to
support this behavior indefinitely. Instead, move the bulk of the script into
a method of CompUnit::RepositoryRegistry. This way we save even a little
compilation time.

Brings down zef's startup from 0.744 to 0.507 seconds.