Speed up finding installed scripts by using a specialized interface
The .files method of CompUnit repositories needs to read the candidate
dists and parse the JSON. When looking for a single script, we are actually
only interested in the path to the file. So use a specialized method that
does only that and can thus avoid parsing any JSON at all. The only
drawback is that for the full gain, the dist has to be re-installed as
previously the short name lookup file did not contain the resource id of
the script.