Fix Mixy (-) Mixy behaviour
(-) is all about subtracting weights.  Since Mixes can have negative
weights, the occurence of a key on the right side that does not exist
on the left side, should cause that key to appear in the result Mix
with a negative weight (as 0 - N = -N).  This also happens with
Baggy (-) Baggy, with the difference being that Baggies cannot have
negative weights.

This breaks one [(^)] test, because either the implementation of
multiple (^) handling is wrong, or the test is wrong.