Re-work Proc so :in($p1.out) will plumb fds.
Previously, we'd do the copying in process. Now this takes advantage
of Proc::Async being able to plumb the file descriptors together. This
does introduce some new complications: we no longer can start to read
output right away when the process is spawned. This means we must work
harder to avoid deadlock in the case input is written before the output
is plumbed into the input of a second process (by buffering up all the
writes and performing them asynchronously).

This change also conflicted with the new behavior of various methods
that started to block until process exit (such as Bool), since they
would hang forever due to reads not having been started. This is
resolved by having them start reading from the output handles if
they were not instead plumbed to something else.