Revert "Stop the optimizer from using what I assume are old semantics for multi dispatch"
This reverts commit ccfa5e51a21f5a613ecb0cfb24be6c15f4103752. The
semantics it removed were not old. They were a special case analysis
of the use of literals in combination with known native types, to
allow native candidates to be called. The reverted commit may have
been intended to fix some inconsistency (which is what I thought was
happening), but in fact the code it removed is critical for good
performance.