Fix a data race in Proc::Async.
It is possible that the ready callback of the process will fire before
the code that binds the async process handle into `$!process_handle`
gets to run. This can happen if the thread that initiates the async
process call is suspended very soon after it has done so, and then
another thread gets the ready message. The result was that the permit
issuing code would run without `$!process_handle` being bound, and so
it would never issue the emit permit, and there would therefore be a
deadlock because the stdout and/or stderr handles were never read.