Make Proc cope with zero-length buffers.
It seems libuv sometimes hands over an empty buffer from stdout or
stderr. These are dutifully passed along, meaning they may also be
observed coming out of Proc::Async. Proc and IO::Pipe assumed that
this would not happen, and mis-handled it doing so as an indication
of the end.