Replace $*MAIN-ALLOW-NAMED-ANYWHERE
With %*SUB-MAIN-OPTS. The way the arguments are parsed is a contentious
issue and we might see more options being proposed. There's a motion[^1]
to have the $*MAIN-ALLOW-NAMED-ANYWHERE be specced, which was mostly
added as a hack[^2] (for panda?). So, pre-emptively let's spec an
options hash in which we can keep this and all future options.

Looking at some of the Getopts::* modules, I think a more pleasant
interface for them could be just hook in via %*SUB-MAIN-OPTS and let
the regular sub MAIN routines accept the args, just like stock Perl 6.

[1] https://github.com/perl6/doc/pull/1397
[2] https://irclog.perlgeek.de/perl6-dev/2017-06-29#i_14803554