Don't force MAIN-SUB-OPTS overhead on every startup
If a developer of a script wants to use MAIN-SUB-OPTS, she will have
to create the lexicaly dynamic herself.  This is in the nature of
dynamic variables: there is no need for the core setting have them
created automagically.  This also sends a bad cognitive signal about
dynamic variables.

This patch also for the time being adds temporary support for the old
MAIN-ALLOW-NAMED-ANYWHERE, so that users of a newer rakudo with an older
panda, will still work.