Fix 'is export' on constants assigned to routines
The 'is export' exception catchall was higher precedence than the
constant exporting candidate in the case of Routine. This patch makes
`constant $sub is export = sub { ... }` take the same path as
`sub foo is export { ... }` while preserving the constant variable's
symbol.

Fixes "Can't use unknown trait 'is export' in a ... declaration" for
statements like:

constant $constant-regex is export =  /fo*/;
constant &constant-sub   is export  = sub { "win" };

Zoffix++ for the patch