Introducing Mixy.total-positive
Intended as an internal method, but may have its uses outside of the
core.  It returns the sum of the positive weights in the Mix(|Hash).
Since this is a fixed value for Mix, we can make Mix.roll quite a bit
faster, especially for Mixes with larger number of elements.