Baggy no longer uses HLL hash internally
Until now, Baggies contained a HLL Hash, which in turn contained a
low-level IterationSet.  Now, Baggies directly have the IterationSet
without the intervening HLL Hash.  This means one less object to
create an allocate for each Baggy created.  This also reduces the
number of calls to Baggy.RAW-HASH, as we can now directly refer to
the underlying IterationSet.