Cache compunits by id instead of spec str
Previously this would cache on a spec string, but the spec
used was the *query* and not a concrete spec.

The earliest a concrete identity can be obtained to cache is
after calling !matching-dist, which handles all the range stuff
so we don't cache something like "6.*" when "6.c" is what will be
loaded. While we could also continue to use the string identity
(just from this concrete version) that would require a Distribution
to be created that is othewise saved by caching on the file-id.