Use implicit bin/ for running scripts
It made sense for the .files/.candidates api to take an explicit
bin/ (if looking for bin/foo). It could make sense for this
.run-script + wrapper, but it also make sense without since the
method name implies the name-path prefix is a bin/ script.

Since a rakudo installation can be upgraded with distributions
(i.e. bin/ scripts) already installed we should continue to use
the current behavior of an implicit bin/ prefix.