Speed up NativeCall subs by compiling specialized subroutine body
Instead of having a generic code block handle all NativeCall subs, we
compile a specialized one for every sub. This way we can actually
utilize the information from the signature to identify the arguments
that need deconting and save the looping over the arguments at runtime.

This also leaves the argument checks to the call dispatcher.