Move supply/react impl classes out of sub.
The class was re-captured every call. This was a bit wasteful. More
notably, however, it led to the class block being lexically captured
by something with a caller chain that eventually led back to the outer
before, meaning that supply blocks might end up building a reference
chain consisting of all the frames involved in processing all of the
prior messages! This could leak a huge amount of memory.