Add atomicint type.
Which will always be of an appropriate size for use with atomic ops
on integers. What that size will be is platform specific. The integer
atomic ops will work on any native int value that happens to be the
current platform's atomic size, but we'll do the documentation so as
to strongly encourage using atomicint instead to achieve portability.