Return Failure from failed P6-level .parse on 6.e
Cherry-pick of @9501edae4f73a970e3270e3b0336a7b3045d3329:

The .parse and .parsefile methods will now return a Failure object that points
to the high-water state of the parser, if available.  (The .subparse method
does not, since it is presumed to be for lower-level piecemeal matching that
is likelier to be success-oriented and have its own failure recovery modes.)

The behavior isn't changing for 6.c and 6.d.

Resolving perl6/6.d-prep#9