Try to produce less containers for $ parameters.
We wrap them up in a Scalar if we can't, at compile-time, see a type
constraint that rules out Iterable. This patch attempts to do that
only if a runtime check determines the parameter to be Iterable, and
just deconts it if not. Unfortunately, this breaks a few tests, and I
don't have time to look into that at the moment. But leaving this here
in case somebody else wishes to have a dig. The branch does mean quite
a bit more generated code on the one hand, but it also means spesh can
toss the appropriate branch, so in many cases tossing out the need to
allocate and assign into a container. I've checked and it does indeed
manage to achieve this - if only the change didn't case regressions.