Grammar.parse to return Nil on fail in 6.c
and a failure in 6.d and higher.  This is achieved by moving all of
the new logic to a new file src/core.d/Grammar.pm which augments the
existing Grammar.pm.  In the original source, the call to .SETFAIL
has become an optional call using .?SET-FAIL, which returns Nil in
case the method doesn't exist (which it won't in 6.c).  Which neatly
matches the original failure value, so no further handling is needed.