Restructure naming of atomic ops
While writing the Perl 6 Weekly, it became clear to me that there is
no simple Unicode -> Texas mapping for many of the new atomic ops.
Since the release is delayed, I took the liberty to make the naming
more consistent:

- atomic-inc became atomic-postfix-inc (texas version of ⚛++)
- atomic-dec became atomic-postfix-dec (texas version if ⚛--)
- atomic-add became atomic-postfix-add (no unicode variant available yet)

- added infix ⚛-= and infix ⚛−= (using U+2212 minus)

- added atomic-prefix-inc as a texas version of ++⚛
- added atomic-prefix-dec as a texas version of --⚛
- added atomic-prefix-add as a texas version of ⚛+=
- added atomic-prefix-sub as a texas version of ⚛-=

- added proto to cas() (how did this ever compile in the setting without one?)

Please note I'm not happy with the naming of prefix/postfix-add/sub, as
the texas naming should reflect whether the returned value is from before
or after the adding/subbing.  But atomic-add-before seems more ambiguous
than atomic-postfix-add, even though we're talking about an infix.