Don't make assumptions about multidim keys bein Int/Str
Multi-dimensional postcircumfix [ ] returned a slice (List) instead
of a scalar element, unless the key at every dimensions was an Int.

(Same with postcircumfix { } and Str, since commit 2faed16.)

That was incorrect. Now, a slice is only returned if any dimension
has an Iterable or Whatever/HyperWhatever key.

Example test-case that now passes:

    my @a = [1, [[2, 3, 4], 5, [6, 7]]];
    is-deeply @a[1; 0.0; "2"],  4, 'indices are coerced to integers';