Implement "bare" REPL and use it on non-TTY uses
This is basically a fake REPL that just evals all the code given
on STDIN and then exits. This lets us support piped usages…

    $ echo -e 'my $x = 42;\nsay $x' | ./perl6
    42

…without REPL user messages being printed. This behavior can be
requested on TTY displays by using RAKUDO_LINE_EDITOR=bare env var.
If the env var is set to one of the other line editors we support,
the bare REPL won't run even if we are on a non-TTY display.