s/index/eqat/g in INTERPOLATE literal handling
The simplification in 215a5fa731 used nqp::index* for the ignore* cases,
but if that change is made to the regular case some spectests break.
Both 215a5fa731 and this change pass all spectests, which I think
indicates that there probably aren't enough (any?) tests for
regex interpolation with ingnore* modifiers.

Using index* is more efficient, since it doesn't require walking
through the haystack and checking at each position, but it somehow needs
to be done more correctly.