Remove unnecessary creation of an extra serialization context
Adding the sub to a newly created SC when it's already added to one
will actively prevent spesh from optimizing as it won't be able to
find the WVal for self. Furthermore this would also prevent us from
JIT compiling the sub's body.

However, adding to an SC is vital when dealing with function pointers
returned from native code as those subs will not be added to an SC by
the compiler.