Enable output buffering by default for non-TTYs
Prior to this change, output buffering had to be explicitly turned
on using `:buffer` when opening a file, and was never turned on for
the standard output handles. Now it will be turned on for handles
that are not TTYs. This gives a sizable performance boost for output.
For example, writing a million line file drops from ~2.05s to ~1.29s,
thus running in 63% of the time it used to. Presuming UTF-8 encoding,
this brings us to being 1.4x slower than Perl 5 (which is also doing
output buffering) at this benchmark, as opposed to 2.2x slower before.