Fix inability of `make`ing type objects
The check for whether we should return a `Nil` from .made/.ast
is meant to detect an NQPMu in $!made. nqp::isconcrete was used
for that purpose, which also makes any `make`d type objects to
result in Nils.

Fix by doing a typecheck for `Mu` that returns false when $!made
is an NQPMu, but would return True for any Rakudo object.