Fix thread safety of "foo{$x}bar"
Previously, we didn't do a closure clone of the block, which meant it
could get its outer captured by a different thread (or go very wrong
if a recursive call was done inside of the closure). Doing this clone
corrects the bug, and brings it in line with other closures, though
does cause a change to `state` inside such as block (again, bringing
it in line with how things usually work, though there was a spectest
based around the previous wrong expectation).