Remove action 13 from BUILDALLPLAN
In BUILDALL this is a noop, so don't put it in the BUILDALLPLAN
to begin with.  And since we now removed that, we don't need to
support this in Mu.BUILDALL either.

This should have some effect minor effect on memory usage, and
maybe some noticeable effect in build times of objects with attributes
that don't need initialization.