Don't start Configure.pl without ExtUtils::Command
Not all distros include this module, but it is needed to run the Makefile.
If a user would try and install Rakudo Star and they don't have the module,
this would obscure the missing module and they likely would not know why
the Rakudo Star compilation/installation was failing.

`use` the module in Configure.pl so the user immediately knows they must
install this module if they do not have it.