Initial re-implementation of thread pool scheduler
Has separate general and timer queues with separate workers, and also
introduces affinity queues, which are intended for cases where events
will be fed into a `Supply`, and thus there's no point having lots of
threads competing over them only to immediately stumble over each
other. The separate timer queue helps with timer events being delayed,
for example by a process producing a load of output.

The new implementation also adds a supervisor, which is where we will
put the smarts on how many threads to have in the pool. For now, it is
already smart enough to start a lot less threads than the previous
scheduler when they obviously aren't needed, which should help a good
bit with memory consumption, as well as to add more when needed to
break deadlocks. This area will be further developed in upcoming
commits. The default maximum number of threads has been raised to 64,
now that the behavior of the scheduler is not to innevitably end up
starting the maximum number of threads even when they don't have any
work to do.

Finally, a new RAKUDO_SCHEDULER_DEBUG environment variable can be set
to see how many and what kinds of threads are being started.