Start factoring CPU cores and usage into scheduler
So that it can pick a better number of threads for the workoad. Now it
only boosts threads beyond the number of CPU cores if detects very low
resource usage in combination with a queue of work, which is suggestive
of a deadlock that may be resolved by an extra thread.