Implement pred() and succ() for the Enumeration role
These method will walk the enumeration in the declaration order.

Using Order as an example:
- Order::Same.succ is Order::More,
- Order::Same.pred is Order::Less.

Calling pred or succ on the boundaries will fail with X::OutOfBound.
Using the same example, Order::Less.pred fails with this X::OutOfRange:
"Decrement out of range. Is: Less, should be in Order::Less^..Order::More".