Fix non-blocking react { await blah() }
The `react` (and also `await $some-supply`) constructs will tap the
`Supply` upon the call to `subscribe-awaiter`. This causes some
amount of synchronous execution to take place, and that code may then
itself `await`. This broke because the continuation tag was removed
by the react/await prior to this. Protect the continuation tag, so it
survives.