Change supply concurrency control mechanism
Previously, a supply/react block would immediately accept all of the
messages sent to it, put them into a queue, and then whatever thread
happened to be processing the current message would help out with
processing new arrivals afterwards. This could cause messages to pile
up without any back-pressure up the chain to slow things down. This
mechanism is also a key part of the situation discussed in RT #130716.

This patch switches over to a different approach, using Lock::Async.
This means that a `supply` block, from the outside, now behaves far
more like everything else in a Supply chain, in that the time it
costs to process the message is paid for by the sender. In the event
of contention, then the current awaiter is used. This means that for
code in the thread pool, an emitter that reaches a supply block that
is busy processing some other message will have a continuation taken,
which will be scheduled once the Supply block is available again.

This change alone would lead to immediate deadlock, however, for any
case where the Supply that is tapped does a synchronous emit or an
`await`, or itself has a `whenever` for anything that does that. As a
minimal example, `react { whenever supply { emit 1 } { .say } }` would
deadlock, because the setup work running the `react` block holds the
lock, and the `emit` in the `supply` block, which is run upon tap,
wants it. It would thus block on the `await`, blocking the completion
of the `react` block body with it, meaning nothing would happen. This
was resolved by the previous queue-then-help mechanism.

Therefore, a special Awaiter is installed when a `whenever` block is
added. The `.tap` call is now done in a block at the root of a
continuation reset. If an `await` happens, a continuation will be
taken. After the processing of either the `supply` or `react` block
mainline has been completed, the continuations will be invoked, with
any further `await`s resulting from that also being invoked, until an
empty list of continuations is reached.

This in turn means that `react { whenever supply { loop { emit 1 } } {
.say } }` will now spit out load of 1s, instead of hanging.

It's not all good news. There is a regression in S17-supply/throttle.t
and the sometimes-failing S17-supply/supplier-preserving.t is now a
reliable failure. The second of these is due to the way tap objects
are provided to the tapper - potentially too late. This also blocks
trying to address the issue of an emitter running synchronously never
terminating (we should at least fire an `on-close` or `CLOSE` block,
and we may be able to make the next `emit` just tear the thing down
too).

The final bit of bad news is that benchmarks on Cro show quite a big
slowdown from this change. It's not entirely clear why, since it's
dealing with real async data sources (sockets), and so should never
need to take a continuation during the setup phase. It's possible
something else needs tuning for this new back-pressure mechanism, or
it's possible there's some performance surprise somewhere in the
changes that needs identifying.