Push Tap objects down to subscribers
The `.tap(...)` call now takes an optional `:&tap` named argument that
will be called, synchronously and ahead of any `emit`/`done`/`quit`,
with the Tap object. This means that there is now a reliable way to
get the Tap object, no matter the behavior of the source Supply. There
were two notable cases where we had trouble:

* A Supply that started to spit out values synchronously would never
  return its Tap object until it was done, so the consumer had no way
  to turn it off
* A Supply that quickly started to emit values on another thread could
  cause a data race. In `react { whenever $quick-emit { done } }`, if
  `$quick-emit` emitted a value before the code in the `react` guts
  had chance to stash away the Tap object, then the call to `.close`
  would be missed. This caused S17-supply/supplier-preserving.t to be
  unreliable and fail now and then.

Both of these are resolved by this commit. It additionally moves the
Tappable implementation classes out of the methods, which found a
couple of bugs and should be a slight speedup.