In slurp, don't call nqp::join if not needed
nqp::join creates a copy of the string(s) during processing and then
creates another copy to return. For slurping large files this can be
very wasteful if there weren't any "\r\n"s to split on.

For a 50mb file of 1million lines, the time to slurp it dropped from
2.3s to 1.8s, and ~230mb of memory didn't need to be allocated.