Add "no queue progress" deadlock heuristic
Will want some further tuning, but this handles the case where we do
not start enough threads because we are working hard on, say, events
fired by the timer queue, but not making any real progress because the
general queue is not making any progress at all. Fixes hang that some
low-core-count users observed in S17-supply/syntax.t.