Start using $*AWAITER in `await` in 6.c
But set an additional flag indicating that we should really block, so
we still get the 6.c semantics, just in terms of the new factoring.
This means that the parts of the new supply concurrency model that
depend on the use of $*AWAITER will now work out with 6.c code that
does an `await` in the mainline of a `supply` block. It also means
that we'll from today be exercising much of the code that makes up
6.d non-blocking await, except the bit in ThreadPoolScheduler that
actually does the non-blocking part of the job.