HLLize shortname parameter to CU::DependencySpecification
This change makes all of the auto-generated BUILDALL stuff work.

Theory is that having a native str as a named parameter doesn't
jive well with doing Str:D checking.  But this can apparently only
happen when being called from NQP, as in Perl 6 a native str will
be hllized automatically.