Compile optimized bodies of native subs at compile time
Previously the optimized bodies of native subs were compiled on the first
call of the sub. NativeCall now exports a CHECK phaser to do this at the
end of the compilation phase. This means that it happens e.g. during
precompilation and can be done once per installation instead of once per run.

Note that loading the native library is still done at runtime. Also since we
only know if we can JIT compile the calling code after nativecallbuild, we
compile both the JITed and non-JITed body candidates and select the appropriate
one in !setup.