Implement IO::Handle.slurp(:bin)
During IO::Grant, we removed :bin from .slurp-rest, because it
mutated the handle's encoding. After the Deasyncing of IO, we
clarified the spec in that it was allowed to use binary read methods
on non-bin handles. With that in mind, re-adding a
non-$!encoding-mutating :bin to .slurp aligns the interface with
other methods and makes it easier to bin-slurp from Proc pipes too.