deconfuse statement modifier errors
Catch missing expressions on statement modifiers earlier so that the
tests eat_terminator are only for statement controls.  Also, move the
cursor to the front of the cursor controls to indicate where the semicolon
is actually perceived to be missing if we report that.  Additionally,
we change the kok check to use .panic instead of .sorry to avoid multiple
errors resulting from repeated terminator checks that use statement
modifier keywords for terminators.

Fixes #125596 and #125674.