Make is-run() mangle run() args on Windows
Due to RT#132258, when we `run $*EXECUTABLE` it goes throw cmd.exe
and the tests fail due to bad quoting. Mangle the args to the
test routine, so they get sent off properly.

The issue with sending double quotes appear to still remain in some
cases and properly working around it seems to need VERBATIM args
send off in libuv, which we currently have off.

Fixes almost all of `make test` failures. The remaining failures just
need to be switched over to this test routine instead of doing `run`
manually, landing onto the same problem.